Changed wording of ion_storm.dm to reflect the fact that NEV Unity is a ship, not a station.
Someone might want to give a quick readthrough to see if I missed any but I changed all instances of "station" into "ship" and flipped the one law where the AI thinks the station is a ship into one where the AI thinks the ship is a station instead.

Plus I added a law about pirates. Because everyone loves pirates.